+++ src/store/index.ts
@@ -1,12 +1,10 @@
 const persistState = require('redux-localstorage');
 const createLogger = require( 'redux-logger');
 import { IAppState, rootReducer, deimmutify, reimmutify } from './store';
-import { ILineup, IParty, LineupActions } from './lineup';
+import { ILineup, IParty } from './lineup';
 import { IMenu, IMenuItem } from './menu';
-import { ITables, ITable, TableActions } from './tables';
+import { ITables, ITable } from './tables';
 import { DevTools } from './dev-tools';
-
-const ACTION_PROVIDERS = [ LineupActions, TableActions ];
 
 export {
   IAppState,
@@ -18,9 +16,6 @@ export {
   ITables,
   ITable,
   DevTools,
-  LineupActions,
-  TableActions,
-  ACTION_PROVIDERS,
 };
 
 export const middleware = [
