+++ src/store/lineup/index.ts
@@ -1,10 +1,8 @@
 import { ILineup, IParty } from './lineup.types';
 import { lineupReducer } from './lineup.reducer';
 import { deimmutifyLineup, reimmutifyLineup } from './lineup.transformers';
-import { LineupActions } from './lineup.actions';
 
 export {
-  LineupActions,
   ILineup,
   IParty,
   lineupReducer,
